Lay the foundations of a Microservice

Generate boiler plate code for a microservice that works out-of-the-box.

use_microservice(
  entrypoint_name = "microservice",
  endpoint_name = "RESTful",
  host = "localhost",
  port = 8080
)

Arguments

entrypoint_name
(character) Entrypoint name.

endpoint_name
(character) Endpoint name.

host
(character) a string that is a valid IPv4 or IPv6 address that is owned by this server, which the application will listen on. "0.0.0.0" represents all IPv4 addresses and "::/0" represents all IPv6 addresses.

port
(integer) a number or integer that indicates the server port that should be listened on. Note that on most Unix-like systems including Linux and Mac OS X, port numbers smaller than 1025 require root privileges.

Value

The necessary files for spinning up a microservice at http://host:port/

Details

What does use_microservice do?

Given entrypoint_name (microservice), endpoint_name (RESTful), host (127.0.0.1) and port number (8080),

When use_microservice is called

use_microservice(
    entrypoint_name = "microservice", 
    endpoint_name = "RESTful",
    host = "127.0.0.1",
    port = 8080             
)

Then the function generates the necessary files for a working-out-of-the-box microservice at http://127.0.0.1:8080/. In particular, the function creates or modifies the following files:

./inst/entrypoints/microservice-foreground.R

./inst/entrypoints/microservice-background.R

./inst/endpoints/RESTful.R

./tests/testthat/test-endpoint-RESTful.R

./tests/testthat/helpers-xyz.R

./tests/testthat/setup-xyz.R

./DESCRIPTION

Spinning up the Microservice

Running the Microservice is done by one of two scripts:

inst/entrypoints/microservice-foreground.R runs the Microservice at the foreground. This option is appropriate during production. However, running the Microservice at this mode means the R session is locked and would not process CLI or other scripts commands until the Microservice is stopped.

inst/entrypoints/microservice-background.Rruns the Microservice at the

background as an RStudio job. This option is appropriate during development. Running the Microservice at this mode allows the programmer to interact with the Microservice by sending it HTTP requests within the same R session.

Discovering the Microservice Details

Communicating with the Microservice is done through HTTP requests. This means you need to know what URL exists, what service they provide, and what input they require.

Firstly, OpenAPI Specification provides visual documentation for the Microservice API. If the Microservice is hosted locally on port 8080, then you can browse OpenAPI at http://127.0.0.1:8080//__docs__/.

Second source of documentation is the Microservice unit-test, located at ./tests/testthat/test-endpoint-RESTful.R. The unit-test fully covers the available services, demonstrates how to send HTTP requests, and provides examples of how to post-process the returned response into an R data.frame.

Developing and Testing Endpoints

use_microservice automatically generates endpoint unit-test that is useful both for developing and testing the API. The boilerplate code of the unit-test looks like:

context("unit test for RESTful endpoint")

# Setup -------------------------------------------------------------------
testthat::setup({
    ## Capture API Information
    # httptest::.mockPaths(file.path(usethis::proj_get(), 'tests', 'api'))
    # httptest::start_capturing()
    # local_mock(with_mock_api = function(expr) eval(expr), .env = "httptest")
    # GET <<- function(...) httptest::capture_requests(httr::GET(...))
    # POST <<- function(...) httptest::capture_requests(httr::POST(...))
})

# healthcheck -------------------------------------------------------------
httptest::with_mock_api({
    test_that("healthcheck returns a success", {
        url <- generate_url(config$host, config$port, "utility/healthcheck")
        expect_status_200(GET(url))
    })
})

# class -------------------------------------------------------------------
httptest::with_mock_api({
    describe("class", {
        it("POST with character returns a character", {
            input <- "Hello World!"
            x <- jsonlite::toJSON(input, auto_unbox = TRUE)
            url <- generate_url(config$host, config$port, "utility/class")
            expect_status_200(response <- POST(url, body = x))

            output <-
                extract_content_text(response) %>%
                jsonlite::fromJSON(flatten = TRUE) %>%
                stringr::str_remove_all('\"')
            expect_match(output, "character")
        })

        it("POST with data.frame returns a data.frame", {
            input <- mtcars
            x <- jsonlite::toJSON(input, auto_unbox = TRUE)
            url <- generate_url(config$host, config$port, "utility/class")
            expect_status_200(response <- POST(url, body = x))

            output <-
                extract_content_text(response) %>%
                jsonlite::fromJSON(flatten = TRUE) %>%
                stringr::str_remove_all('\"')
            expect_match(output, "data.frame")
        })
    })
})

# Teardown ----------------------------------------------------------------
testthat::teardown({httptest::stop_capturing()})

See also