Add a Workflow that Models a Domain

Create a workflow script with sequential steps.

use_workflow(name, domain, n_step = 3)

Arguments

name
(character) Workflow name. Note: workflow applies naming convention automatically.

domain
(character) sub-domain name. Note: workflow applies naming convention automatically.

n_step
(character) Number of steps that constitute the Workflow.

Details

Highlights

Create a workflow in R that is readable by a domain expert.

Facilitates the transition between concept to implementation of the domain model.

What does use_workflow do?

Given the workflow name (order one pizza), the sub-domain name (pizza ordering) and the number of steps in the workflow (3),

When use_workflow is called

use_workflow(name = "order one pizza", domain = "pizza ordering", n_step = 3)

Then the function:

Creates a workflow boilerplate with three steps;

Places the file at inst/workflows/order-one-pizza.R; and

Opens the file (only in interactive mode).

What is a Step

The input to a step should always be an environment.

The output of a step should always be an environment.

session <- new.env

step <- function(session){
  
  session$system <- list(date = Sys.Date(), os = Sys.info()["sysname"])

  return(session)
}

How to Move from Conception to Implementation?

In this section we customise the boilerplate code to fit the application design.

Iteration 0 - Decompose a Process into Steps

Before you call use_workflow, think about the process you are about to model. Essentially, you need to decompose the process into sequential steps that reflect their relationships and dependencies.

Decomposing a process into steps requires cognitive effort, yet you don’t need to delve into details in this stage.

Tip: Aim to capure 20% of the steps that are roughly 80% of the process.

Note: Even the best planning is not so omniscient as to get it right the first time.

Caution: Each step may (but don’t have to) be dependent on its predecessors steps.Each step must be independent from its succeeding steps.

In our running example, the process of ordering pizza includes, but not limited to, three steps:

Issue new order;

Order pizza; and

Commit order.

These three necessary steps for ordering pizza happen sequentially in the order of their writing. In this case, their order is irreversible, each step is dependent on its predecessors steps.

We start the implementation process by generating boilerplate code for a workflow that comprises three steps:

The boilerplate code produced by use_workflow looks likes this:

################################################################################
## order-one-pizza Workflow
################################################################################
pkgload::load_all()

# Steps -------------------------------------------------------------------
step_1 <- function(session){
    return(session)
}
step_2 <- function(session){
    return(session)
}
step_3 <- function(session){
    return(session)
}

# Workflow ----------------------------------------------------------------
session <- new.env()

session %>%
    step_1() %>%
    step_2() %>%
    step_3()

Iteration 1 - Rename the Steps

In this iteration we rename the generic names of the steps. Each step name, should reveal you why the step exists, what it does and how it is used.

################################################################################
## order-one-pizza Workflow
################################################################################
pkgload::load_all()

# Steps -------------------------------------------------------------------
issue_new_order <- function(session){
    return(session)
}
order_pizza <- function(session){
    return(session)
}
commit_order <- function(session){
    return(session)
}

# Workflow ----------------------------------------------------------------
session <- new.env()

session %>%
    issue_new_order() %>%
    order_pizza() %>%
    commit_order()

Iteration 2 - ???

Tip: hello

See also